requirements.txt
mysql-connector-python 
contextlib3
pytest